+++ tests/cell_test.py
@@ -285,10 +285,10 @@ def test_get_polygons_depth(tree):
 
 def test_get_polygons_filter(tree):
     c3, c2, c1 = tree
-    polys = c3.get_polygons(layer=3)
-    assert len(polys) == 12
-    polys = c3.get_polygons(datatype=3)
-    assert len(polys) == 12
+    with pytest.raises(ValueError):
+        _ = c3.get_polygons(layer=3)
+    with pytest.raises(ValueError):
+        _ = c3.get_polygons(datatype=3)
     polys = c3.get_polygons(layer=0, datatype=0)
     assert len(polys) == 6
     polys = c3.get_polygons(layer=1, datatype=1)